+++ swing/framework-ui/src/main/java/is/codion/swing/framework/ui/EntityTableCellRenderer.java
@@ -24,6 +24,7 @@ import is.codion.framework.domain.entity.EntityDefinition;
 import is.codion.framework.domain.entity.attribute.Attribute;
 import is.codion.framework.domain.entity.attribute.AttributeDefinition;
 import is.codion.swing.common.model.component.table.FilterTableModel;
+import is.codion.swing.common.ui.component.table.FilterTable;
 import is.codion.swing.common.ui.component.table.FilterTableCellRenderer;
 import is.codion.swing.common.ui.component.table.FilterTableCellRenderer.ColorProvider;
 import is.codion.swing.framework.model.SwingEntityTableModel;
@@ -71,7 +72,11 @@ public final class EntityTableCellRenderer {
 	/**
 	 * @param <T> the attribute value type
 	 */
-	public interface EntityColorProvider<T> extends ColorProvider<Entity, Attribute<?>, T> {}
+	public interface EntityColorProvider<T> extends ColorProvider<Entity, Attribute<?>, T> {
+
+		@Override
+		Color color(FilterTable<Entity, Attribute<?>> table, Entity entity, Attribute<?> attribute, T value);
+	}
 
 	/**
 	 * A {@link SwingEntityTableModel} based table cell factory.