+++ changelog.md
@@ -46,6 +46,7 @@ Codion Change Log
 - FilterTableCellRenderer.Builder row and column identifier type parameters added.
 - FilterTable.Builder.cellEditor() and cellRenderer() now long Supplier based.
 - FilterTableCellRenderer, renaming and refactoring.
+- FilterTableCellRenderer.ColorProvider.color() row and column index parameters replaced with row item and column identifier.
 ### is.codion.framework.domain
 - Entity.copy(), copyBuilder() and deepCopy() moved to Entity.Copy, renamed mutable(), builder() and deep() respectively, immutable() added.
 - ImmutableEntity bug fixed, StackOverflowError when creating an immutable entity with cyclical foreign key dependencies, some refactoring.